Rename scripts to snake_case and update paths
