+++ scripts/prompt_to_text.py
